refactor: T MAP API에서 사용하는 데이터 클래스의 패키지 정리
