--- data/src/main/java/com/stop/data/remote/model/route/Currency.kt
@@ -1,12 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class Currency(
-    @Json(name = "currency")
-    val currency: String,
-    @Json(name = "currencyCode")
-    val currencyCode: String,
-    @Json(name = "symbol")
-    val symbol: String
-)
