--- data/src/main/java/com/stop/data/remote/model/route/Fare.kt
@@ -1,8 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class Fare(
-    @Json(name = "regular")
-    val regular: Regular
-)
