--- data/src/main/java/com/stop/data/remote/model/route/PassShape.kt
@@ -1,8 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class PassShape(
-    @Json(name = "linestring")
-    val linestring: String
-)
