--- data/src/main/java/com/stop/data/remote/model/route/PassStopList.kt
@@ -1,8 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class PassStopList(
-    @Json(name = "stationList")
-    val stationList: List<Station>
-)
