--- data/src/main/java/com/stop/data/remote/model/route/Plan.kt
@@ -1,8 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class Plan(
-    @Json(name = "itineraries")
-    val itineraries: List<Itinerary>
-)
