--- data/src/main/java/com/stop/data/remote/model/route/Regular.kt
@@ -1,10 +1,0 @@
-package com.stop.data.remote.model.route
-
-import com.squareup.moshi.Json
-
-data class Regular(
-    @Json(name = "currency")
-    val currency: Currency,
-    @Json(name = "totalFare")
-    val totalFare: Int
-)
