--- domain/src/main/java/com/stop/domain/model/RouteResponse.kt
@@ -1,5 +1,0 @@
-package com.stop.domain.model
-
-data class RouteResponse (
-    val itineraries: List<Itinerary>
-)
