+++ domain/src/main/java/com/stop/domain/model/route/tmap/RouteRequest.kt
@@ -1,0 +1,11 @@
+package com.stop.domain.model.route.tmap
+
+data class RouteRequest(
+    val startX: String,
+    val startY: String,
+    val endX: String,
+    val endY: String,
+    val lang: Int = 0,
+    val format: String = "json",
+    val count: Int = 10,
+)
