+++ domain/src/main/java/com/stop/domain/model/route/tmap/custom/Station.kt
@@ -1,8 +1,9 @@
-package com.stop.domain.model
+package com.stop.domain.model.route.tmap.custom
 
 data class Station(
     val orderIndex: Int,
     val coordinate: Coordinate,
     val stationId: String,
     val stationName: String,
+    val idUsedAtPublicApi: String,
 )
