+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/Leg.kt
@@ -1,0 +1,18 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class Leg(
+    val distance: Double,
+    val end: End,
+    val mode: String,
+    val passShape: PassShape?,
+    val passStopList: PassStopList?,
+    val route: String?,
+    val routeColor: String?,
+    val sectionTime: Double,
+    val start: Start,
+    val steps: List<Step>?,
+    val type: Int?
+)
