+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/MetaData.kt
@@ -1,0 +1,9 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MetaData(
+    val plan: Plan,
+    val requestParameters: RequestParameters
+)
