+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/PassShape.kt
@@ -1,0 +1,8 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class PassShape(
+    val linestring: String
+)
