+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/PassStopList.kt
@@ -1,0 +1,8 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class PassStopList(
+    val stationList: List<Station>
+)
