+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/Regular.kt
@@ -1,0 +1,9 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class Regular(
+    val currency: Currency,
+    val totalFare: Int
+)
