+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/Start.kt
@@ -1,0 +1,10 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class Start(
+    val lat: Double,
+    val lon: Double,
+    val name: String
+)
