+++ domain/src/main/java/com/stop/domain/model/route/tmap/origin/Step.kt
@@ -1,0 +1,11 @@
+package com.stop.domain.model.route.tmap.origin
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class Step(
+    val description: String,
+    val distance: Int,
+    val linestring: String,
+    val streetName: String
+)
