+++ domain/src/main/java/com/stop/domain/usecase/GetRouteUseCase.kt
@@ -1,9 +1,9 @@
 package com.stop.domain.usecase
 
-import com.stop.domain.model.RouteRequest
-import com.stop.domain.model.RouteResponse
+import com.stop.domain.model.route.tmap.custom.Itinerary
+import com.stop.domain.model.route.tmap.RouteRequest
 
 interface GetRouteUseCase {
 
-    suspend fun getRoute(routeRequest: RouteRequest): RouteResponse
+    suspend fun getRoute(routeRequest: RouteRequest): List<Itinerary>
 }
